requirements.txt
Django==5.0.2
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
django-environ==0.11.2
django-filter==23.5
django-allauth==0.61.0
django-ckeditor==6.7.0
django-cleanup==8.1.0
psycopg2-binary==2.9.9
python-slugify==8.0.4
whitenoise==6.6.0
gunicorn==21.2.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
django-rest-auth==0.9.5
django-leaflet==0.29.0
geocoder==1.38.1
channels==4.0.0
channels-redis==4.2.0
redis==5.0.1
celery==5.3.6 